Reorganization into the correct project file structure for a python package.
