+++ ecoliframalpha/variability_metrics.csv
@@ -1,0 +1,5 @@
+variance,Fano_factor,CV,CRI,codon
+0.04691918141892219,0.06558927596212123,0.30280122200263815,1.0999912188604688,AAA
+0.045906427958032694,0.06484894223566216,0.3026677484867762,1.0895935557190244,GAT
+0.0280494524892712,0.13690443632005975,0.8174389495006015,0.4169939138969561,CGT
+0.02674075615402051,0.13360486999833412,0.817025190548995,0.41270058887963623,CTG
